balancing
health

+++ Assets/Scripts/UI/HealthBar.cs
@@ -15,7 +15,7 @@ public class HealthBar : MonoBehaviour
         healthBar.maxValue = playerHealth.maxHealth;
         healthBar.value = playerHealth.maxHealth;
     }
-    public void SetHealth(int hp)
+    public void SetHealth(float hp)
     {
         healthBar.maxValue = playerHealth.maxHealth;
         healthBar.value = hp;
